Remove Typeorm files, rename some files for coherency and adapt the tsconfig.json 'include'
