+++ apps/api/src/data-administrator-context/adapters/secondary/gateways/repositories/drizzle/sql-nomination-file.repository.it-spec.ts
@@ -9,7 +9,7 @@ import {
   DrizzleDb,
   getDrizzleInstance,
 } from 'src/shared-kernel/adapters/secondary/repositories/drizzle/drizzle-instance';
-import { DrizzleTransactionPerformer } from 'src/shared-kernel/adapters/secondary/providers/drizzleTransactionPerformer';
+import { DrizzleTransactionPerformer } from 'src/shared-kernel/adapters/secondary/providers/drizzle-transaction-performer';
 import { nominationFiles } from './schema/nomination-file-pm';
 import { drizzleConfigForTest } from 'src/shared-kernel/adapters/secondary/repositories/drizzle/drizzle-config';
