+++ apps/api/src/reporter-context/adapters/secondary/gateways/repositories/drizzle/sql-nomination-file-report.repository.ts
@@ -1,7 +1,7 @@
 import { eq } from 'drizzle-orm';
 import { ReportRepository } from 'src/reporter-context/business-logic/gateways/repositories/report.repository';
 import { NominationFileReport } from 'src/reporter-context/business-logic/models/nomination-file-report';
-import { DrizzleTransactionableAsync } from 'src/shared-kernel/adapters/secondary/providers/drizzleTransactionPerformer';
+import { DrizzleTransactionableAsync } from 'src/shared-kernel/adapters/secondary/providers/drizzle-transaction-performer';
 import { buildConflictUpdateColumns } from 'src/shared-kernel/adapters/secondary/repositories/drizzle/drizzle-sql-preparation';
 import { DateOnly } from 'src/shared-kernel/business-logic/models/date-only';
 import { reports } from './schema/report-pm';
