+++ apps/api/src/reporter-context/business-logic/use-cases/report-update/updateReport.spec.ts
@@ -1,5 +1,5 @@
 import { FakeNominationFileReportRepository } from 'src/reporter-context/adapters/secondary/gateways/repositories/fake-nomination-file-report.repository';
-import { NullTransactionPerformer } from 'src/shared-kernel/adapters/secondary/providers/nullTransactionPerformer';
+import { NullTransactionPerformer } from 'src/shared-kernel/adapters/secondary/providers/null-transaction-performer';
 import { TransactionPerformer } from 'src/shared-kernel/business-logic/gateways/providers/transactionPerformer';
 import { NominationFileReport } from '../../models/nomination-file-report';
 import { ReportBuilder } from '../../models/report.builder';
