--- apps/api/src/shared-kernel/adapters/secondary/repositories/orm-data-source.ts
@@ -1,4 +1,0 @@
-import { DataSource } from 'typeorm';
-import { ormConfig } from './orm-config';
-
-export default new DataSource(ormConfig('src'));
