--- apps/api/src/shared-kernel/adapters/secondary/repositories/typeorm/dateTransformers.ts
@@ -1,10 +1,0 @@
-import { ValueTransformer } from 'typeorm';
-
-export class DateTransformer implements ValueTransformer {
-  to(value: Date): Date {
-    return value;
-  }
-  from(value: string): Date {
-    return new Date(value);
-  }
-}
