+++ apps/api/tsconfig.json
@@ -53,6 +53,6 @@
       }
     ],
   },
-  "include": ["src/**/*.ts", "drizzle.config.ts"],
+  "include": ["src/**/*.ts","test/**/*.ts", "drizzle.config.ts"],
   "references":[{"path": "../../packages/shared-models"}],
 }
